+++ index.js
@@ -1,11 +1,34 @@
+var rc = require("rc");
+var log = require("loglevel");
+
 exports.SUCCESS = "SUCCESS";
 exports.FAILED = "FAILED";
 
-exports.send = function(event, context, responseStatus, responseData, physicalResourceId) {
+
+
+exports.send = function(event, context, responseStatus, responseData, physicalResourceId, options) {
+
+  // rc mutates defaults, so we define them here
+  var DEFAULTS = {
+    returnError: true,
+    logLevel: "info"
+  };
+
+  options = options || {};
+
+  var cfg = rc("cfn_responder",DEFAULTS,options);
+  var doneCallback = context.done;
+  log.setLevel(cfg.logLevel);
 
   responseData = responseData || {};
   if (typeof responseData !== 'object') {
-    return context.done("'Data' must be a key/pair object");
+    if (cfg.returnError) {
+      return doneCallback("'Data' must be a key/pair object");
+    }
+    else {
+      responseData = {};
+      log.info("'Data' was set to an empty object because it was not a valid key/pair object");
+    }
   }
 
   var jsonBody = {
@@ -25,13 +48,14 @@ exports.send = function(event, context, responseStatus, responseData, physicalRe
     jsonBody.PhysicalResourceId = exports.FAILED;
   }
 
+  log.debug(jsonBody);
   var responseBody = JSON.stringify(jsonBody);
 
   var https = require("https");
   var url = require("url");
 
   var parsedUrl = url.parse(event.ResponseURL);
-  var options = {
+  var httpOptions = {
     hostname: parsedUrl.hostname,
     port: 443,
     path: parsedUrl.path,
@@ -41,16 +65,31 @@ exports.send = function(event, context, responseStatus, responseData, physicalRe
       "content-length": responseBody.length
     }
   };
+  log.debug("httpOptions",httpOptions);
 
-  var request = https.request(options, function(response) {
+  var request = https.request(httpOptions, function(response) {
     if (response.statusCode === 200)
-      return context.done(null, jsonBody);
+      return doneCallback(null, jsonBody);
 
-    context.done("Failed to communicate to S3: " + response.statusCode,jsonBody);
+    var errMsg = "Failed to communicate to S3: " + response.statusCode;
+    if (cfg.returnError) {
+      return doneCallback(errMsg,jsonBody);
+    }
+    else {
+      log.info(errMsg);
+      return doneCallback(null);
+    }
   });
 
   request.on("error", function(error) {
-    context.done("send(..) failed executing https.request(..): " + error);
+    var errMsg = "send(..) failed executing https.request(..): " + error;
+    if (cfg.returnError) {
+      doneCallback(errMsg);
+    }
+    else {
+      log.info(errMsg);
+      doneCallback(null);
+    }
   });
 
   request.write(responseBody);