+++ test/responder.test.js
@@ -34,6 +34,22 @@ describe("cfn-responder", function() {
     responder.send(event,context,responder.SUCCESS);
   });
 
+  it("should return failure if no communication to S3", function(cb) {
+
+    var event = {
+      ResponseURL: "https://fake2.url"
+    };
+
+    var context = {
+      done: function(err,obj) {
+        assert.ifError(err);
+        cb();
+      }
+    };
+    responder.send(event,context,responder.SUCCESS,null,null,{returnError: false});
+  });
+
+
   it("should contain a PhysicalResourceId if the RequestType is Create and status is failed", function(cb) {
 
     nock('https://fake.url')
@@ -112,5 +128,31 @@ describe("cfn-responder", function() {
     };
     responder.send(event,context,responder.SUCCESS,{},1);
   });
+
+  it("should return success if a non-200 is return from S3 and returnError is false", function(cb) {
+
+    nock('https://fake.url')
+    .put('/', {"Status":"SUCCESS","Reason":"See the details in CloudWatch Log Stream: undefined","StackId":"arn:aws:cloudformation:us-east-1:namespace:stack/stack-name/guid","RequestId":"unique id for this create request","LogicalResourceId":"name of resource in template","Data":{}})
+    .reply(500, {});
+
+    var event = {
+      ResponseURL: "https://fake.url",
+      RequestType: "Create",
+      RequestId: "unique id for this create request",
+      ResponseURL: "https://fake.url",
+      ResourceType: "Custom::MyCustomResourceType",
+      LogicalResourceId: "name of resource in template",
+      StackId: "arn:aws:cloudformation:us-east-1:namespace:stack/stack-name/guid"
+    };
+
+    var context = {
+      done: function(err,obj) {
+        assert.ifError(err);
+        cb();
+      }
+    };
+    responder.send(event,context,responder.SUCCESS,{},1,{returnError: false});
+  });
+
 
 });